hot fixes

+++ src/modules/users/services/UpdateUserAvatarServices.ts
@@ -17,26 +17,20 @@ class UpdateUserAvatarServices {
 
 		const user = await userRepository.findById(user_id);
 
-		
-
 		if (!user) {
 			throw new AppError('User not found');
 		}
 
-/* 		if (user.avatar) {
-			const userAvatarFilePath =  path.join(uploadConfig.directory, user.avatar);
+		if (user.avatar) {
+			const userAvatarFilePath = path.join(uploadConfig.directory, user.avatar);
 			const userAvatarFileExists = await fs.promises.stat(userAvatarFilePath);
 			if (userAvatarFileExists) {
 				await fs.promises.unlink(userAvatarFilePath);
 			}
-		} */
-		user.avatar =  avatarFilename;
-	
+		}
+		user.avatar = avatarFilename;
 
-
-	 await userRepository.save(user);
-
-		
+		await userRepository.save(user);
 
 		return user;
 	}
